Refactor public-api tests into integration tests
Generally, things that can be done with only the public API ought to be
tested there, partly to exercise and demonstrate that the public API is
enough.
